+++ docs/changelog.md
@@ -1,9 +1,13 @@
 ## Unreleased
 
-## New in 2.7.2 (2024-11-12)
+## New in 2.7.3 (2024-11-20)
 
 - Listing actions now will include `query` and `predict` decorators.
 - Add `Configure Action Input` internal command.
+
+
+## New in 2.7.2 (2024-11-12)
+
 - Set `SEMA4AI_CREDENTIAL_API` environment variable if port is communicated from Studio
 - Add packages and tools information when submitting an issue.
 - Update Action Server to `2.0.0`.